requirements.txt
altgraph==0.17.2
boto3==1.19.1
botocore==1.22.1
certifi==2021.10.8
click==8.0.3
configparser==5.0.2
contextlib2==21.6.0
Cython==0.29.24
h5py==3.1.0
imageio==2.9.0
jmespath==0.10.0
macholib==1.15.2
mccabe==0.6.1
mypy-extensions==0.4.3
nnabla==1.22.0
numpy==1.21.3
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.0
pycodestyle==2.8.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2021.10.8
s3transfer==0.5.0
scipy==1.7.1
six==1.16.0
tomli==1.2.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
